fix(cluster helper func): double quoting in keyspace names
When keyspace names start with a digit, they are returned with
surrounding double quotes. This can lead to double quoting in CQL
statements that can lead to SCT failure. This commit updates the
get_test_keyspaces func to strip these extra quotes. If keyspace
isnt quoted the change is harmless